Created DiskOutputTreeTest and tested everything that is written and loaded back
