+++ tdlbcpp/src/BUILD
@@ -7,7 +7,7 @@ cc_library(
 cc_library(
     name = "computeUnit",
     hdrs = ["ComputeUnit.h", "Header.h", "Output.hpp", "DiskOutputTree.h", "Boundary.hpp", "CollisionEntropic.hpp", "GlobalStructures.hpp", "StreamingEsoteric.hpp", "cxxopts.hpp", "CollisionEgglesSomers.hpp", "ComputeUnit.hpp", "Forcing.hpp", "QVec.hpp", "StreamingNieve.hpp"],
-    visibility = ["//tdlbcpp/tests/Params:__pkg__"],
+    visibility = ["//tdlbcpp/tests/Params:__pkg__", "//tdlbcpp/tests:__pkg__"],
 )
 
 cc_binary(
