+++ tdlbcpp/tests/Params/BUILD
@@ -1,3 +1,15 @@
+cc_library(
+    name = "ParamsCommon",
+    hdrs = ["ParamsCommon.hpp"],
+    srcs = ["ParamsCommon.cpp"],
+    visibility = ["//tdlbcpp/tests/Params:__pkg__", "//tdlbcpp/tests:__pkg__"],
+    deps = ["//tdlbcpp/src/Params:params",
+            "//tdlbcpp/src:computeUnit",
+            "//tdlbcpp/tests:utils", 
+            "@googletest//:gtest_main"],
+    copts = ["-Itdlbcpp/src"]
+)
+
 cc_test(
   name = "tests",
   srcs = ["BinFileTest.cpp", "CheckpointTest.cpp", "ComputeUnitParamsTest.cpp", "OutputParamsTest.cpp", "FlowTest.cpp", "GridTest.cpp", "RunningTest.cpp"],
@@ -5,6 +17,7 @@ cc_test(
           "//tdlbcpp/src/Params:params",
           "//tdlbcpp/src:computeUnit",
           "//tdlbcpp/tests:utils",
+          "//tdlbcpp/tests/Params:ParamsCommon",
           "@googletest//:gtest_main"
   ],
   copts = ["-Itdlbcpp/src"]
