+++ tdlbcpp/tests/Params/BinFileTest.cpp
@@ -14,25 +14,12 @@
 #include "Params/BinFile.hpp"
 
 #include "tdlbcpp/tests/utils.hpp"
+#include "tdlbcpp/tests/Params/ParamsCommon.hpp"
 
 class BinFileTests : public ::testing::Test
 {
 protected:
     std::string filename;
-
-    void checkAllFields(BinFileParams &expected, BinFileParams &actual)
-    {
-        ASSERT_EQ(expected.filePath, actual.filePath) << "filePath field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.name, actual.name) << "name field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.note, actual.note) << "note field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.structName, actual.structName) << "structName field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.binFileSizeInStructs, actual.binFileSizeInStructs) << "binFileSizeInStructs field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.coordsType, actual.coordsType) << "coordsType field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.hasGridtCoords, actual.hasGridtCoords) << "hasGridtCoords field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.hasColRowtCoords, actual.hasColRowtCoords) << "hasColRowtCoords field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.QDataType, actual.QDataType) << "QDataType field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.QOutputLength, actual.QOutputLength) << "QOutputLength field has a wrong value after being written to a file and then read";
-    }
 
 public:
     BinFileTests()
@@ -47,17 +34,7 @@ public:
 
 TEST_F(BinFileTests, BinFileWriteReadValidTest)
 {
-    BinFileParams binFileParams;
-    binFileParams.filePath = "test";
-    binFileParams.name = "test1";
-    binFileParams.note = "test2";
-    binFileParams.structName = "test3";
-    binFileParams.binFileSizeInStructs = 2;
-    binFileParams.coordsType = "test4";
-    binFileParams.hasGridtCoords = true;
-    binFileParams.hasColRowtCoords = true;
-    binFileParams.QDataType = "test5";
-    binFileParams.QOutputLength = 0x8fffffff;
+    BinFileParams binFileParams = ParamsCommon::createBinFileParamsFixed();
 
     binFileParams.writeParamsToJsonFile(filename);
     std::cerr << filename << std::endl;
@@ -65,22 +42,12 @@ TEST_F(BinFileTests, BinFileWriteReadValidTest)
     BinFileParams binFileParamsRead;
     binFileParamsRead.getParamsFromJsonFile(filename);
 
-    checkAllFields(binFileParamsRead, binFileParams);
+    ParamsCommon::checkAllFields(binFileParamsRead, binFileParams);
 }
 
 TEST_F(BinFileTests, BinFileWriteReadValidRandomTest)
 {
-    BinFileParams binFileParams;
-    binFileParams.filePath = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.name = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.note = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.structName = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.binFileSizeInStructs = rand();
-    binFileParams.coordsType = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.hasGridtCoords = (rand() & 1) == 1;
-    binFileParams.hasColRowtCoords = (rand() & 1) == 1;
-    binFileParams.QDataType = TestUtils::random_string(TestUtils::randomStringLength);
-    binFileParams.QOutputLength = rand();
+    BinFileParams binFileParams = ParamsCommon::createBinFileParamsRandom();
 
     binFileParams.writeParamsToJsonFile(filename);
     std::cerr << filename << std::endl;
@@ -88,7 +55,7 @@ TEST_F(BinFileTests, BinFileWriteReadValidRandomTest)
     BinFileParams binFileParamsRead;
     binFileParamsRead.getParamsFromJsonFile(filename);
 
-    checkAllFields(binFileParamsRead, binFileParams);
+    ParamsCommon::checkAllFields(binFileParamsRead, binFileParams);
 }
 
 TEST_F(BinFileTests, BinFileParamsReadInValidTest)
