+++ tdlbcpp/tests/Params/FlowTest.cpp
@@ -7,7 +7,6 @@
 
 #include <cstdio>
 #include <cstdlib>
-#include <random>
 
 #include "gtest/gtest.h"
 
@@ -15,58 +14,12 @@
 #include "Params/Flow.hpp"
 
 #include "tdlbcpp/tests/utils.hpp"
+#include "tdlbcpp/tests/Params/ParamsCommon.hpp"
 
 class FlowParamsTests : public ::testing::Test
 {
 protected:
     std::string filename;
-
-    template <typename T>
-    FlowParams<T> createFlowParamsWithRandomValues()
-    {
-        FlowParams<T> flowParams;
-
-        std::uniform_real_distribution<double> unif(TestUtils::randomDoubleLowerBound, TestUtils::randomDoubleUpperBound);
-        std::default_random_engine re;
-
-        flowParams.initialRho = (T)unif(re);
-        flowParams.reMNonDimensional = (T)unif(re);
-        flowParams.uav = (T)unif(re);
-        flowParams.cs0 = (T)unif(re);
-        flowParams.g3 = (T)unif(re);
-        flowParams.nu = (T)unif(re);
-        flowParams.fx0 = (T)unif(re);
-        flowParams.Re_m = (T)unif(re);
-        flowParams.Re_f = (T)unif(re);
-        flowParams.uf = (T)unif(re);
-        flowParams.alpha = (T)unif(re);
-        flowParams.beta = (T)unif(re);
-        flowParams.useLES = (rand() & 1) == 1;
-        flowParams.collision = TestUtils::random_string(TestUtils::randomStringLength);
-        flowParams.streaming = TestUtils::random_string(TestUtils::randomStringLength);
-
-        return flowParams;
-    }
-
-    template <typename T>
-    void checkAllFields(FlowParams<T> &expected, FlowParams<T> &actual)
-    {
-        ASSERT_EQ(expected.initialRho, actual.initialRho) << "initialRho field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.reMNonDimensional, actual.reMNonDimensional) << "reMNonDimensional field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.uav, actual.uav) << "uav field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.cs0, actual.cs0) << "cs0 field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.g3, actual.g3) << "g3 field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.nu, actual.nu) << "nu field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.fx0, actual.fx0) << "fx0 field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.Re_m, actual.Re_m) << "Re_m field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.Re_f, actual.Re_f) << "Re_f field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.uf, actual.uf) << "uf field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.alpha, actual.alpha) << "alpha field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.beta, actual.beta) << "beta field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.useLES, actual.useLES) << "useLES field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.collision, actual.collision) << "collision field has a wrong value after being written to a file and then read";
-        ASSERT_EQ(expected.streaming, actual.streaming) << "streaming field has a wrong value after being written to a file and then read";
-    }
 
 public:
     FlowParamsTests()
@@ -81,23 +34,7 @@ public:
 
 TEST_F(FlowParamsTests, FlowDoubleWriteReadValidTest)
 {
-    FlowParams<double> flowParams;
-
-    flowParams.initialRho = 0.0;
-    flowParams.reMNonDimensional = 0.1;
-    flowParams.uav = 0.2;
-    flowParams.cs0 = 0.3;
-    flowParams.g3 = 0.4;
-    flowParams.nu = 0.5;
-    flowParams.fx0 = 0.6;
-    flowParams.Re_m = 0.7;
-    flowParams.Re_f = 0.8;
-    flowParams.uf = 0.9;
-    flowParams.alpha = 1.0;
-    flowParams.beta = 1.1;
-    flowParams.useLES = true;
-    flowParams.collision = "test1";
-    flowParams.streaming = "test2";
+    FlowParams<double> flowParams = ParamsCommon::createFlowParamsFixed();
 
     flowParams.writeParamsToJsonFile(filename);
     std::cerr << filename << std::endl;
@@ -105,12 +42,12 @@ TEST_F(FlowParamsTests, FlowDoubleWriteReadValidTest)
     FlowParams<double> flowParamsRead;
     flowParamsRead.getParamsFromJsonFile(filename);
 
-    checkAllFields(flowParams, flowParamsRead);
+    ParamsCommon::checkAllFields(flowParams, flowParamsRead);
 }
 
 TEST_F(FlowParamsTests, FlowDoubleWriteReadRandomValidTest)
 {
-    FlowParams<double> flowParams = createFlowParamsWithRandomValues<double>();
+    FlowParams<double> flowParams = ParamsCommon::createFlowParamsWithRandomValues<double>();
 
     flowParams.writeParamsToJsonFile(filename);
     std::cerr << filename << std::endl;
@@ -118,12 +55,12 @@ TEST_F(FlowParamsTests, FlowDoubleWriteReadRandomValidTest)
     FlowParams<double> flowParamsRead;
     flowParamsRead.getParamsFromJsonFile(filename);
 
-    checkAllFields(flowParams, flowParamsRead);
+    ParamsCommon::checkAllFields(flowParams, flowParamsRead);
 }
 
 TEST_F(FlowParamsTests, FlowFloatWriteReadRandomValidTest)
 {
-    FlowParams<float> flowParams = createFlowParamsWithRandomValues<float>();
+    FlowParams<float> flowParams = ParamsCommon::createFlowParamsWithRandomValues<float>();
 
     flowParams.writeParamsToJsonFile(filename);
     std::cerr << filename << std::endl;
@@ -131,7 +68,7 @@ TEST_F(FlowParamsTests, FlowFloatWriteReadRandomValidTest)
     FlowParams<float> flowParamsRead;
     flowParamsRead.getParamsFromJsonFile(filename);
 
-    checkAllFields(flowParams, flowParamsRead);
+    ParamsCommon::checkAllFields(flowParams, flowParamsRead);
 }
 
 TEST_F(FlowParamsTests, FlowParamsReadInValidTest)
