+++ tdlbcpp/tests/utils.hpp
@@ -1,3 +1,6 @@
+#ifndef UTILS_H_INCLUDED
+#define UTILS_H_INCLUDED
+
 #include <string>
 #include <cstdlib>
 
@@ -10,11 +13,15 @@ namespace TestUtils
     const int randomArraySize = 5;
     const int randomArrayMinimalSize = 2;
 
-    const std::string temporaryFilenameSuffx = "_to_delete.json";
+    const std::string temporaryFilenameSuffix = "_to_delete.json";
 
     std::string random_string(size_t length);
     std::string getTempFilename();
+    std::string getTempFilename(std::string folder, std::string suffix);
     void removeTempFile(const std::string filename);
     void captureStderr();
     std::string getCapturedStderr();
+    std::string joinPath(const std::string &p1, const std::string &p2);
 }
+
+#endif //UTILS_H_INCLUDED
